+++ ArmPkg/Include/Library/ArmStandaloneMmCoreEntryPoint.h
