+++ ArmPkg/Library/ArmStandaloneMmCoreEntryPoint/ArmStandaloneMmCoreEntryPoint.inf
@@ -1,7 +1,7 @@
 ## @file
 # Module entry point library for DXE core.
 #
-# Copyright (c) 2017 - 2021, Arm Ltd. All rights reserved.<BR>
+# Copyright (c) 2024, Arm Ltd. All rights reserved.<BR>
 #
 #  SPDX-License-Identifier: BSD-2-Clause-Patent
 #
@@ -10,33 +10,56 @@
 
 [Defines]
   INF_VERSION                    = 0x0001001A
-  BASE_NAME                      = StandaloneMmCoreEntryPoint
-  FILE_GUID                      = C97AC593-109A-4C63-905C-675FDE2689E8
+  BASE_NAME                      = ArmStandaloneMmCoreEntryPoint
+  FILE_GUID                      = 16d7b2e4-a025-11ef-8931-6b032fa329a6
   MODULE_TYPE                    = MM_CORE_STANDALONE
   VERSION_STRING                 = 1.0
   PI_SPECIFICATION_VERSION       = 0x00010032
   LIBRARY_CLASS                  = StandaloneMmCoreEntryPoint|MM_CORE_STANDALONE
 
 #
-#  VALID_ARCHITECTURES           = IA32 X64 IPF EBC (EBC is for build only)
+#  VALID_ARCHITECTURES           = ARM AARCH64
 #
 
-[Sources.X64]
-  X64/StandaloneMmCoreEntryPoint.c
+[Sources]
+  ArmStandaloneMmCoreEntryPoint.c
+  SetPermissions.c
+
+[Sources.AARCH64]
+  AArch64/ModuleEntryPoint.S
+
+[Sources.ARM]
+  Arm/ModuleEntryPoint.S
 
 [Packages]
+  ArmPkg/ArmPkg.dec
   MdePkg/MdePkg.dec
   MdeModulePkg/MdeModulePkg.dec
   StandaloneMmPkg/StandaloneMmPkg.dec
 
+[Packages]
+  ArmPkg/ArmPkg.dec
+
 [LibraryClasses]
   BaseLib
   DebugLib
+  StandaloneMmMmuLib
+  ArmSvcLib
+  ArmTransferListLib
+  ArmFfaLib
 
 [Guids]
   gMpInformationHobGuid
   gEfiMmPeiMmramMemoryReserveGuid
   gEfiStandaloneMmNonSecureBufferGuid
+  gEfiHobListGuid
+
+[Protocols]
+  gEdkiiPiMmCpuDriverEpProtocolGuid
+  gEfiMmCommunication2ProtocolGuid
+
+[Pcd]
+  gArmTokenSpaceGuid.PcdStMmStackSize
 
 #
 # This configuration fails for CLANGPDB, which does not support PIE in the GCC
